applying minor fixups to juno barbican spec
This change makes a number of cleanup related fixes, and moves the
-api.service file to directly use wsgi instead of calling the
barbican.sh script.

It also removes the patches applied to excise pbr and fix the
barbican.sh file. The former is related to the downstream relaxation on
pbr and the latter is based on the barbican.sh not being used.

* changing api service file to use uwsgi directly
* adding pbr to build requires
* adding uwsgi and python plugin to requires
* specifically including vassals
* changing install step to skip rebuilding the python package
* changing files to specifically include barbican python files
* removing pbr patch
* removing barbican.sh patch